Add VERY BASIC group hierarchy nesting hack
This first go at describing the group hierarchy (by nesting sub-groups
under their parents, and providing a button on the parent group row to
show/hide child rows), is just a PoC and is more to start a discussion
around design and how to itterate on improvement, than an attempt at
final implementation. From this basic prototype, an idea of the intended
look & feel can be gleaned, and a discussion around what else (e.g. user
usage counts) could/should be nested is now possible.